+++ internal/core/items.rs
@@ -691,7 +691,7 @@ declare_item_vtable! {
 }
 
 declare_item_vtable! {
-    fn slint_get_PinchGestureHandlerVTable() -> PinchGestureHandlerVTable for PinchGestureHandler
+    fn slint_get_ScaleRotateGestureHandlerVTable() -> ScaleRotateGestureHandlerVTable for ScaleRotateGestureHandler
 }
 
 #[repr(C)]